+++ src/main/java/io/github/astrapi69/time/convert/OffsetDateTimeExtensions.java
@@ -1,8 +1,8 @@
 /**
  * The MIT License
- *
+ * <p>
  * Copyright (C) 2015 Asterios Raptis
- *
+ * <p>
  * Permission is hereby granted, free of charge, to any person obtaining
  * a copy of this software and associated documentation files (the
  * "Software"), to deal in the Software without restriction, including
@@ -10,10 +10,10 @@
  * distribute, sublicense, and/or sell copies of the Software, and to
  * permit persons to whom the Software is furnished to do so, subject to
  * the following conditions:
- *
+ * <p>
  * The above copyright notice and this permission notice shall be
  * included in all copies or substantial portions of the Software.
- *
+ * <p>
  * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND,
  * EXPRESS OR IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF
  * MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND
@@ -24,8 +24,11 @@
  */
 package io.github.astrapi69.time.convert;
 
+import java.sql.Timestamp;
+import java.time.LocalDateTime;
 import java.time.OffsetDateTime;
 import java.time.ZoneId;
+import java.time.ZoneOffset;
 import java.time.ZonedDateTime;
 import java.util.Date;
 import java.util.Objects;
@@ -77,6 +80,73 @@ public class OffsetDateTimeExtensions
 	{
 		Objects.requireNonNull(date);
 		Objects.requireNonNull(zone);
-		return date.toInstant().atZone(zone).toOffsetDateTime();
+		return toOffsetDateTime(date, zone.getRules().getOffset(date.toInstant()));
 	}
+
+	/**
+	 * Converts the given {@link Date} with the given {@link ZoneOffset} to a {@link OffsetDateTime}
+	 * object
+	 *
+	 * @param date
+	 *            the date
+	 * @param zoneOffset
+	 *            the zone offset
+	 * @return The {@link OffsetDateTime} object
+	 */
+	public static OffsetDateTime toOffsetDateTime(Date date, ZoneOffset zoneOffset)
+	{
+		Objects.requireNonNull(date);
+		Objects.requireNonNull(zoneOffset);
+		return date.toInstant().atOffset(zoneOffset);
+	}
+
+	/**
+	 * Converts the given {@link LocalDateTime} with the given {@link ZoneId} to a
+	 * {@link OffsetDateTime} object
+	 *
+	 * @param localDateTime
+	 *            the date
+	 * @param zone
+	 *            the zone id
+	 * @return The {@link OffsetDateTime} object
+	 */
+	public static OffsetDateTime toOffsetDateTime(LocalDateTime localDateTime, ZoneId zone)
+	{
+		Objects.requireNonNull(localDateTime);
+		Objects.requireNonNull(zone);
+		return localDateTime.atZone(zone).toOffsetDateTime();
+	}
+
+	/**
+	 * Converts the given {@link LocalDateTime} with the given {@link ZoneOffset} to a
+	 * {@link OffsetDateTime} object
+	 *
+	 * @param localDateTime
+	 *            the date
+	 * @param zoneOffset
+	 *            the zone offset
+	 * @return The {@link OffsetDateTime} object
+	 */
+	public static OffsetDateTime toOffsetDateTime(LocalDateTime localDateTime,
+		ZoneOffset zoneOffset)
+	{
+		Objects.requireNonNull(localDateTime);
+		Objects.requireNonNull(zoneOffset);
+		return localDateTime.atOffset(zoneOffset);
+	}
+
+	/**
+	 * Converts the given {@link OffsetDateTime} to a {@link Timestamp} object
+	 *
+	 * @param offsetDateTime
+	 *            the {@link OffsetDateTime} object
+	 * @return The {@link ZonedDateTime} object
+	 */
+	public static Timestamp toTimestamp(OffsetDateTime offsetDateTime)
+	{
+		Objects.requireNonNull(offsetDateTime);
+		LocalDateTime localDateTime = offsetDateTime.toLocalDateTime();
+		return Timestamp.valueOf(localDateTime);
+	}
+
 }