+++ src/main/java/io/github/astrapi69/time/period/model/Period.java
@@ -1,8 +1,8 @@
 /**
  * The MIT License
- *
+ * <p>
  * Copyright (C) 2015 Asterios Raptis
- *
+ * <p>
  * Permission is hereby granted, free of charge, to any person obtaining
  * a copy of this software and associated documentation files (the
  * "Software"), to deal in the Software without restriction, including
@@ -10,10 +10,10 @@
  * distribute, sublicense, and/or sell copies of the Software, and to
  * permit persons to whom the Software is furnished to do so, subject to
  * the following conditions:
- *
+ * <p>
  * The above copyright notice and this permission notice shall be
  * included in all copies or substantial portions of the Software.
- *
+ * <p>
  * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND,
  * EXPRESS OR IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF
  * MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND
@@ -52,7 +52,7 @@ public class Period
 
 	/**
 	 * Factory method for create a new {@link Period} object
-	 * 
+	 *
 	 * @param start
 	 *            The start {@link Instant} object
 	 * @param end
@@ -155,7 +155,7 @@ public class Period
 
 	/**
 	 * Gets the start {@link Period} object
-	 * 
+	 *
 	 * @return the start {@link Period} object
 	 */
 	public Instant getStart()
@@ -165,7 +165,7 @@ public class Period
 
 	/**
 	 * Gets the end {@link Period} object
-	 * 
+	 *
 	 * @return the end {@link Period} object
 	 */
 	public Instant getEnd()